dom/indexedDB/test/test_getAll.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const description = "My Test Database";

      const values = [ "a", "1", 1, "foo", 300, true, false, 4.5, null ];

      let request = moz_indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let objectStore = db.createObjectStore("foo", "", true);

      request = objectStore.getAll();
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 0, "No elements");

      let addedCount = 0;

      for (let i in values) {
        request = objectStore.add(values[i]);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          if (++addedCount == values.length) {
            SimpleTest.executeSoon(function() { testGenerator.next(); });
          }
        }
      }
      yield;

      request = db.transaction("foo").objectStore("foo").getAll();
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, values.length, "Same length");

      for (let i in event.result) {
        is(event.result[i], values[i], "Same value");
      }

      request = db.transaction("foo").objectStore("foo").getAll(null, 5);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 5, "Correct length");

      for (let i in event.result) {
        is(event.result[i], values[i], "Same value");
      }

      let keyRange = moz_indexedDB.makeBoundKeyRange(1, 9);

      request = db.transaction("foo").objectStore("foo").getAll(keyRange);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, values.length, "Correct length");

      for (let i in event.result) {
        is(event.result[i], values[i], "Same value");
      }

      keyRange = moz_indexedDB.makeBoundKeyRange(4, 7);

      request = db.transaction("foo").objectStore("foo").getAll(keyRange);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 4, "Correct length");

      for (let i in event.result) {
        is(event.result[i], values[parseInt(i) + 3], "Same value");
      }

      keyRange = moz_indexedDB.makeBoundKeyRange(4, 7);

      request = db.transaction("foo").objectStore("foo").getAll(keyRange, 2);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 2, "Correct length");

      for (let i in event.result) {
        is(event.result[i], values[parseInt(i) + 3], "Same value");
      }

      keyRange = moz_indexedDB.makeBoundKeyRange(4, 7);

      request = db.transaction("foo").objectStore("foo").getAll(keyRange, 50);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 4, "Correct length");

      for (let i in event.result) {
        is(event.result[i], values[parseInt(i) + 3], "Same value");
      }

      keyRange = moz_indexedDB.makeBoundKeyRange(4, 7);

      request = db.transaction("foo").objectStore("foo").getAll(keyRange, 0);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 0, "Correct length");

      keyRange = moz_indexedDB.makeBoundKeyRange(4, 7, true, true);

      request = db.transaction("foo").objectStore("foo").getAll(keyRange);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      is(event.result instanceof Array, true, "Got an array object");
      is(event.result.length, 2, "Correct length");

      for (let i in event.result) {
        is(event.result[i], values[parseInt(i) + 4], "Same value");
      }

      finishTest();
      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
